cell list and hash table implemented as derived types
This commit brings type encapsulation to the hash table and the cell list. This
means that we can now handle several cell lists in one run. Usually we don't
need this, but if we want to support cross-set verlet lists then two cell lists
must be built in the same time.

Code compiles and test cases succeed.

git-svn-id: https://ppm.inf.ethz.ch/svn/ppmcore/trunk/libppm@1100 7c7fe9aa-52eb-4d9e-b0a8-ba7d787348e9