A working VTK particle output routine for ppm
Fixed the previously committed code and added a small test case
for checking (this test case is a bit stupid because it requires
the user to actually check the output in paraview - but hey, it's
better than nothing).
This code work with old-style ppm data structures. No particle data
structure assumed, instead a simple derived type is provided that
the user can set up to point at xp and friends for the vtk_particles
routine to find the data to print.