made alloc routines a bit leaner
the substart and substop are now only compiled when the __DEBUG flag is
defined. When running in production (or non-debugging) mode those routines
should be as fast as possible, also when not really doing debugging, the
ppm_debug>0 should not spit out so much output that the really interesting
timings (and stack trace) get lost in the flood of ppm_alloc messages.