re-introducing the GPU branch
When moving from weeman to the ppm server somehow the GPU branch got
destroyed. This commit is remerging all the code introduced in this branch.

Unfortunately the commit history has to stay in a seperate folder (we are
not actually merging, but just adding all the code manually in the branch)